Remove datagen_old package
