+++ src/main/java/org/moddingx/libx/impl/datagen/loot/LootData.java
@@ -1,4 +1,4 @@
-package org.moddingx.libx.impl.datagen_old;
+package org.moddingx.libx.impl.datagen.loot;
 
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.storage.loot.entries.EmptyLootItem;
