+++ src/main/java/org/moddingx/libx/impl/datagen/patchouli/content/EntityContent.java
@@ -1,4 +1,4 @@
-package org.moddingx.libx.impl.datagen_old.patchouli.content;
+package org.moddingx.libx.impl.datagen.patchouli.content;
 
 import com.google.gson.JsonObject;
 import net.minecraft.world.entity.EntityType;
