+++ src/main/java/org/moddingx/libx/impl/datagen/patchouli/content/ImageContent.java
@@ -1,4 +1,4 @@
-package org.moddingx.libx.impl.datagen_old.patchouli.content;
+package org.moddingx.libx.impl.datagen.patchouli.content;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
