+++ src/main/java/org/moddingx/libx/impl/datagen/patchouli/content/MultiblockContent.java
@@ -1,4 +1,4 @@
-package org.moddingx.libx.impl.datagen_old.patchouli.content;
+package org.moddingx.libx.impl.datagen.patchouli.content;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
